+++ src/main/java/ar/edu/unlam/nuralign/domain/models/Report.java
@@ -1,5 +1,6 @@
-/*package ar.edu.unlam.nuralign.domain.models;
+package ar.edu.unlam.nuralign.domain.models;
 
+import jakarta.persistence.Column;
 import lombok.*;
 
 import java.time.LocalDate;
@@ -20,15 +21,16 @@ public class Report {
 
     private LocalDate effectiveDate;
 
-    private Boolean mooodFlag;
+    private Boolean moodFlag;
 
     private Boolean sleepFlag;
 
     private Boolean therapyFlag;
 
-    private Boolean MedicationFlag;
+    private Boolean medicationFlag;
 
-    private String PatientName;
+    private String patientName;
+
+    private Long patientId;
 
 }
-*/